CM-512: Standard and Custom Numeric Format Strings in .NET
