+++ strings-csharp/StandardAndCustomFormatStrings/StandardAndCustomFormatStrings.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StandardAndCustomFormatStrings", "StandardAndCustomFormatStrings\StandardAndCustomFormatStrings.csproj", "{E994D855-2249-4DDF-B58C-ED2BD2CDEEA2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StandardAndCustomFormatStringsTests", "StandardAndCustomFormatStringsTests\StandardAndCustomFormatStringsTests.csproj", "{940BD1E5-A221-42E6-88EC-AC24D6C9F87C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E994D855-2249-4DDF-B58C-ED2BD2CDEEA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E994D855-2249-4DDF-B58C-ED2BD2CDEEA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E994D855-2249-4DDF-B58C-ED2BD2CDEEA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E994D855-2249-4DDF-B58C-ED2BD2CDEEA2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{940BD1E5-A221-42E6-88EC-AC24D6C9F87C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{940BD1E5-A221-42E6-88EC-AC24D6C9F87C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{940BD1E5-A221-42E6-88EC-AC24D6C9F87C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{940BD1E5-A221-42E6-88EC-AC24D6C9F87C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
